pyproject.toml
[project]
name = "dust3r-pytorch-inference-minimal"
version = "0.1.0"
description = "Minimal code for running the Dust3r model with PyTorch"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "einops>=0.8.1",
    "huggingface_hub>=0.29.3",
    "rerun-sdk>=0.22.1",
    "opencv-python>=4.11.0",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "imread-from-url>=0.1.3",
]

[project.optional-dependencies]
onnx = [
    "onnx>=1.17.0",
    "onnxruntime>=1.21.0",
    "onnx-simplifier>=0.4.36",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu121", marker = "sys_platform != 'darwin'" },
]
torchvision = [
    { index = "pytorch-cu121", marker = "sys_platform != 'darwin'" },
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true